# docs/swagger.yaml
basePath: /
definitions:
  dto.CommentCreateDTO:
    properties:
      text:
        maxLength: 1000
        minLength: 1
        type: string
    required:
    - text
    type: object
  dto.CommentResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      post_id:
        type: string
      text:
        type: string
      user_id:
        type: string
    type: object
  dto.DirectMessageCreateDTO:
    properties:
      text:
        maxLength: 1000
        minLength: 1
        type: string
    required:
    - text
    type: object
  dto.PostCreateDTO:
    properties:
      text:
        maxLength: 1000
        minLength: 1
        type: string
    required:
    - text
    type: object
  dto.PostResponse:
    properties:
      author_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      lokesCount:
        type: integer
      text:
        type: string
      updated_at:
        type: string
    type: object
  dto.UserAuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.UserLoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.UserRegisterDTO:
    properties:
      email:
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.UserResponse:
    properties:
      bio:
        type: string
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  model.Conversation:
    properties:
      created_at:
        type: string
      id:
        type: string
      messages:
        items:
          $ref: '#/definitions/model.DirectMessage'
        type: array
      updated_at:
        type: string
      user_a_id:
        type: string
      user_b_id:
        type: string
    type: object
  model.DirectMessage:
    properties:
      conversation_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      sender_id:
        type: string
      text:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@petmessenger.com
    name: API Support
    url: http://www.petmessenger.com/support
  description: This is the backend API for the Pet Messenger social app.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Pet Messenger API
  version: "1.0"
paths:
  /comments/{id}:
    delete:
      description: Delete a comment by ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete comment
      tags:
      - comments
  /conversations:
    get:
      description: Get all conversations of authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Conversation'
            type: array
      summary: List conversations
      tags:
      - conversations
    post:
      consumes:
      - application/json
      description: Create a new conversation with a target user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Conversation'
      summary: Start conversation
      tags:
      - conversations
  /conversations/{id}/messages:
    get:
      description: Get all messages in a conversation
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DirectMessage'
            type: array
      summary: Get messages
      tags:
      - conversations
    post:
      consumes:
      - application/json
      description: Send a message in a conversation
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: Message content
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/dto.DirectMessageCreateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DirectMessage'
      summary: Send message
      tags:
      - conversations
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login data
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserAuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /logout:
    post:
      description: Revoke all refresh tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout user
      tags:
      - auth
  /posts:
    post:
      consumes:
      - application/json
      description: Create a new post for the authenticated user
      parameters:
      - description: Post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/dto.PostCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PostResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a post
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Delete a post by the authenticated user
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete post
      tags:
      - posts
    get:
      description: Get post details
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PostResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get post by ID
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update a post by the authenticated user
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/dto.PostCreateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PostResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update post
      tags:
      - posts
  /posts/{id}/comments:
    get:
      description: Get all comments for a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CommentResponse'
            type: array
      summary: Get comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Add a comment to a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/dto.CommentCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommentResponse'
      summary: Create comment
      tags:
      - comments
  /posts/{id}/like:
    post:
      description: Like a post by the authenticated user
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Like post
      tags:
      - posts
  /posts/{id}/unlike:
    post:
      description: Remove like from a post by the authenticated user
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Unlike post
      tags:
      - posts
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user account with username, email and password
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegisterDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Users
  /users/{id}/posts:
    get:
      description: Get all posts created by a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PostResponse'
            type: array
      summary: Get posts by user
      tags:
      - posts
swagger: "2.0"